+++ kythe/cxx/common/kzip_writer_test.cc
@@ -22,6 +22,7 @@
 #include <unordered_map>
 #include <unordered_set>
 
+#include "absl/status/status.h"
 #include "absl/strings/str_cat.h"
 #include "absl/strings/str_replace.h"
 #include "absl/strings/string_view.h"
@@ -44,7 +45,7 @@ absl::string_view TestTmpdir() {
 }
 
 std::string TestFile(absl::string_view basename) {
-  return absl::StrCat(TestSourceRoot(), "kythe/cxx/common/testdata/",
+  return absl::StrCat(TestSourceRoot(), "kythe/testdata/platform/",
                       absl::StripPrefix(basename, "/"));
 }
 
@@ -55,12 +56,12 @@ struct WithStatusFn {
     return status->ok();
   }
 
-  Status* status;
+  absl::Status* status;
   T function;
 };
 
 template <typename T>
-WithStatusFn<T> WithStatus(Status* status, T function) {
+WithStatusFn<T> WithStatus(absl::Status* status, T function) {
   return WithStatusFn<T>{status, std::move(function)};
 }
 
@@ -75,30 +76,31 @@ std::string TestOutputFile(absl::string_view basename) {
 
 StatusOr<std::unordered_map<std::string, std::unordered_set<std::string>>>
 CopyIndex(IndexReader* reader, IndexWriter* writer) {
-  Status error;
+  absl::Status error;
   std::unordered_map<std::string, std::unordered_set<std::string>> digests;
-  Status scan = reader->Scan(WithStatus(&error, [&](absl::string_view digest) {
-    auto unit = reader->ReadUnit(digest);
-    if (!unit.ok()) {
-      return unit.status();
-    }
-    auto written_digest = writer->WriteUnit(*unit);
-    if (!written_digest.ok()) {
-      return written_digest.status();
-    }
-    for (const auto& file : unit->unit().required_input()) {
-      auto data = reader->ReadFile(file.info().digest());
-      if (!data.ok()) {
-        return data.status();
-      }
-      auto written_file = writer->WriteFile(*data);
-      if (!written_file.ok()) {
-        return written_file.status();
-      }
-      digests[*written_digest].insert(*written_file);
-    }
-    return OkStatus();
-  }));
+  absl::Status scan =
+      reader->Scan(WithStatus(&error, [&](absl::string_view digest) {
+        auto unit = reader->ReadUnit(digest);
+        if (!unit.ok()) {
+          return unit.status();
+        }
+        auto written_digest = writer->WriteUnit(*unit);
+        if (!written_digest.ok()) {
+          return written_digest.status();
+        }
+        for (const auto& file : unit->unit().required_input()) {
+          auto data = reader->ReadFile(file.info().digest());
+          if (!data.ok()) {
+            return data.status();
+          }
+          auto written_file = writer->WriteFile(*data);
+          if (!written_file.ok()) {
+            return written_file.status();
+          }
+          digests[*written_digest].insert(*written_file);
+        }
+        return absl::OkStatus();
+      }));
   if (!scan.ok()) {
     return scan;
   }
@@ -110,18 +112,19 @@ CopyIndex(IndexReader* reader, IndexWriter* writer) {
 
 StatusOr<std::unordered_map<std::string, std::unordered_set<std::string>>>
 ReadDigests(IndexReader* reader) {
-  Status error;
+  absl::Status error;
   std::unordered_map<std::string, std::unordered_set<std::string>> digests;
-  Status scan = reader->Scan(WithStatus(&error, [&](absl::string_view digest) {
-    auto unit = reader->ReadUnit(digest);
-    if (!unit.ok()) {
-      return unit.status();
-    }
-    for (const auto& file : unit->unit().required_input()) {
-      digests[std::string(digest)].insert(file.info().digest());
-    }
-    return OkStatus();
-  }));
+  absl::Status scan =
+      reader->Scan(WithStatus(&error, [&](absl::string_view digest) {
+        auto unit = reader->ReadUnit(digest);
+        if (!unit.ok()) {
+          return unit.status();
+        }
+        for (const auto& file : unit->unit().required_input()) {
+          digests[std::string(digest)].insert(file.info().digest());
+        }
+        return absl::OkStatus();
+      }));
   if (!scan.ok()) {
     return scan;
   }