+++ kythe/java/com/google/devtools/kythe/extractors/java/JavaCompilationUnitExtractor.java
@@ -78,6 +78,7 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.EnumSet;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -287,12 +288,23 @@ public class JavaCompilationUnitExtractor {
       List<String> sourceFiles,
       String outputPath) {
     CompilationUnit.Builder unit = CompilationUnit.newBuilder();
-    unit.setVName(VName.newBuilder().setSignature(target).setLanguage("java"));
+    unit.getVNameBuilder().setSignature(target).setLanguage("java");
     unit.addAllArgument(options);
     unit.setHasCompileErrors(hasErrors);
     unit.addAllRequiredInput(requiredInputs);
+    ImmutableMap<String, String> inputCorpus =
+        Streams.stream(requiredInputs)
+            .collect(
+                ImmutableMap.toImmutableMap(
+                    f -> f.getInfo().getPath(), f -> f.getVName().getCorpus()));
+    Set<String> sourceFileCorpora = new HashSet<>();
     for (String sourceFile : sourceFiles) {
       unit.addSourceFile(sourceFile);
+      sourceFileCorpora.add(inputCorpus.getOrDefault(sourceFile, ""));
+    }
+    if (sourceFileCorpora.size() == 1) {
+      // Attribute the source files' corpus to the CompilationUnit if it is unambiguous.
+      unit.getVNameBuilder().setCorpus(Iterables.getOnlyElement(sourceFileCorpora));
     }
     unit.setOutputKey(outputPath);
     unit.setWorkingDirectory(stableRoot(rootDirectory, options, requiredInputs));
@@ -339,7 +351,6 @@ public class JavaCompilationUnitExtractor {
       Iterable<String> sourcepath,
       Iterable<String> processorpath,
       Iterable<String> processors,
-      Optional<Path> genSrcDir,
       Iterable<String> options,
       String outputPath)
       throws ExtractionException {
@@ -350,7 +361,6 @@ public class JavaCompilationUnitExtractor {
     Preconditions.checkNotNull(sourcepath);
     Preconditions.checkNotNull(processorpath);
     Preconditions.checkNotNull(processors);
-    Preconditions.checkNotNull(genSrcDir);
     Preconditions.checkNotNull(options);
     Preconditions.checkNotNull(outputPath);
     return extract(
@@ -362,7 +372,6 @@ public class JavaCompilationUnitExtractor {
             StandardLocation.SOURCE_PATH, sourcepath,
             StandardLocation.ANNOTATION_PROCESSOR_PATH, processorpath),
         processors,
-        genSrcDir,
         options,
         outputPath);
   }
@@ -373,7 +382,6 @@ public class JavaCompilationUnitExtractor {
       Iterable<String> sources,
       Map<Location, Iterable<String>> searchPaths,
       Iterable<String> processors,
-      Optional<Path> genSrcDir,
       Iterable<String> options,
       String outputPath)
       throws ExtractionException {
@@ -381,15 +389,13 @@ public class JavaCompilationUnitExtractor {
     Preconditions.checkNotNull(sources);
     Preconditions.checkNotNull(searchPaths);
     Preconditions.checkNotNull(processors);
-    Preconditions.checkNotNull(genSrcDir);
     Preconditions.checkNotNull(options);
     Preconditions.checkNotNull(outputPath);
 
     final AnalysisResults results =
         Iterables.isEmpty(sources)
             ? new AnalysisResults()
-            : runJavaAnalysisToExtractCompilationDetails(
-                sources, searchPaths, processors, genSrcDir, options);
+            : runJavaAnalysisToExtractCompilationDetails(sources, searchPaths, processors, options);
 
     List<FileData> fileContents = ExtractorUtils.convertBytesToFileDatas(results.fileContents);
     List<FileInput> compilationFileInputs =
@@ -565,16 +571,22 @@ public class JavaCompilationUnitExtractor {
   private void findRequiredFiles(
       UsageAsInputReportingFileManager fileManager,
       Map<URI, String> sourceFiles,
-      Optional<Path> genSrcDir,
       AnalysisResults results)
       throws ExtractionException {
+    Set<String> genSrcDirs =
+        fileManager.hasLocation(StandardLocation.SOURCE_OUTPUT)
+            ? ImmutableSet.copyOf(
+                Iterables.transform(
+                    fileManager.getLocation(StandardLocation.SOURCE_OUTPUT),
+                    f -> ExtractorUtils.tryMakeRelative(rootDirectory, f.toString())))
+            : ImmutableSet.of();
     for (InputUsageRecord input : fileManager.getUsages()) {
       processRequiredInput(
           input.fileObject(),
           input.location().orElse(null),
           fileManager,
           sourceFiles,
-          genSrcDir,
+          genSrcDirs,
           results);
     }
   }
@@ -584,7 +596,7 @@ public class JavaCompilationUnitExtractor {
       Location location,
       UsageAsInputReportingFileManager fileManager,
       Map<URI, String> sourceFiles,
-      Optional<Path> genSrcDir,
+      Set<String> genSrcDirs,
       AnalysisResults results)
       throws ExtractionException {
     URI uri = requiredInput.toUri();
@@ -695,13 +707,17 @@ public class JavaCompilationUnitExtractor {
       }
     }
 
-    // Identify generated sources by checking if the source file is under the genSrcDir.
-    if (genSrcDir.isPresent()
-        && !isJarPath
-        && requiredInput.getKind() == Kind.SOURCE
-        && Paths.get(relativePath).startsWith(genSrcDir.get())) {
-      results.explicitSources.add(strippedPath);
-      results.newSourcePath.add(genSrcDir.get().toString());
+    // Identify generated sources by checking if the source file is under the genSrcDirs.
+    try {
+      if (!genSrcDirs.isEmpty()
+          && !isJarPath
+          && requiredInput.getKind() == Kind.SOURCE
+          && fileManager.contains(StandardLocation.SOURCE_OUTPUT, requiredInput)) {
+        results.explicitSources.add(strippedPath);
+        results.newSourcePath.addAll(genSrcDirs);
+      }
+    } catch (IOException err) {
+      // Ignore IOExceptions potentially thrown by fileManager.contains(...);
     }
 
     if (!results.fileContents.containsKey(strippedPath)) {
@@ -822,7 +838,6 @@ public class JavaCompilationUnitExtractor {
       Iterable<String> sources,
       Map<Location, Iterable<String>> searchPaths,
       Iterable<String> processors,
-      Optional<Path> genSrcDir,
       Iterable<String> options)
       throws ExtractionException {
 
@@ -835,11 +850,6 @@ public class JavaCompilationUnitExtractor {
       }
 
       results.hasErrors = !task.compile(options, sources, processors);
-
-      // Ensure generated source directory is relative to root.
-      genSrcDir =
-          genSrcDir.map(
-              p -> Paths.get(ExtractorUtils.tryMakeRelative(rootDirectory, p.toString())));
 
       for (String source : sources) {
         results.explicitSources.add(ExtractorUtils.tryMakeRelative(rootDirectory, source));
@@ -855,8 +865,7 @@ public class JavaCompilationUnitExtractor {
       addSystemFiles(results);
 
       // We accumulate all file contents from the java compiler.
-      findRequiredFiles(
-          task.getFileManager(), mapClassesToSources(task.getSymbolTable()), genSrcDir, results);
+      findRequiredFiles(task.getFileManager(), mapClassesToSources(task.getSymbolTable()), results);
     }
 
     return results;